split up utils in message package
