+++ packages/message/package.json
@@ -50,6 +50,7 @@
   },
   "devDependencies": {
     "ava": "^2.1.0",
+    "crypto-random-string": "^3.2.0",
     "nyc": "^14.0.0",
     "rewire": "^4.0.1"
   }
